NoneDiscountPolicy와 DiscountPolicy의 개념적 결합을 분리한다
