+++ src/main/java/com/namu/object/_02_movie/AmountDiscountPolicy.java
@@ -1,6 +1,6 @@
 package com.namu.object._02_movie;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
     private Money discountAmount;
 
     public AmountDiscountPolicy(Money discountAmount, DiscountCondition... conditions) {
