+++ src/main/java/com/namu/object/_02_movie/DiscountPolicy.java
@@ -1,24 +1,6 @@
 package com.namu.object._02_movie;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
-public abstract class DiscountPolicy {
-    private List<DiscountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DiscountCondition ... conditions) {
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening) {
-        for (DiscountCondition each : conditions) {
-            if (each.isSatisfiedBy(screening)) {
-                return getDiscountAmount(screening);
-            }
-        }
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
+public interface DiscountPolicy {
+    Money calculateDiscountAmount(Screening screening);
 }
