+++ src/main/java/com/namu/object/_02_movie/NoneDiscountPolicy.java
@@ -2,7 +2,7 @@ package com.namu.object._02_movie;
 
 public class NoneDiscountPolicy extends DiscountPolicy {
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
