+++ src/main/java/com/namu/object/_02_movie/PercentDiscountPolicy.java
@@ -1,6 +1,6 @@
 package com.namu.object._02_movie;
 
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
